+++ R/mydata.R
@@ -34,6 +34,7 @@
 ##'
 ##' The same data is also provide in \code{'.csv'} format via the openair
 ##'   project web site \url{http://www.openair-project.org}.
+##' @export
 ##' @keywords datasets
 ##' @examples
 ##'