+++ R/trajLevel.R
@@ -299,6 +299,16 @@ trajLevel <- function(mydata, lon = "lon", lat = "lat",
         statistic = "XX" ## i.e. it wont touch the data
     }
 
+    ## location of receptor for map projection, used to show location on maps
+    origin <- head(subset(mydata, hour.inc == 0), 1) ## origin
+    tmp <- mapproject(x = origin[["lon"]][1],
+                      y = origin[["lat"]][1],
+                      projection = projection,
+                      parameters = parameters,
+                      orientation = orientation)
+    receptor <- c(tmp$x, tmp$y)
+    
+
     if (method == "hexbin") {
 
         ## transform data for map projection
@@ -441,7 +451,8 @@ trajLevel <- function(mydata, lon = "lon", lat = "lat",
                              map.cols = map.cols, map.alpha = map.alpha, traj = TRUE,
                              projection = projection,
                              parameters = parameters, orientation = orientation,
-                             grid.col = grid.col, trajLims = trajLims)
+                             grid.col = grid.col, trajLims = trajLims,
+                             receptor = receptor)
 
     ## reset for Args
     scatterPlot.args <- listUpdate(scatterPlot.args, Args)